modularized room controller
